requirements.txt
asgiref==3.2.10
attrs==20.2.0
beautifulsoup4==4.9.1
boto3==1.14.60
botocore==1.17.60
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
Collectfast==2.2.0
cssselect==1.1.0
dj-database-url==0.5.0
Django==3.1.1
django-debug-toolbar==2.2
django-filter==2.3.0
django-s3-folder-storage==0.5
django-storages==1.10.1
djangorestframework==3.11.1
docutils==0.15.2
drf-nested-routers==0.91
fire==0.3.1
gunicorn==20.0.4
idna==2.10
iniconfig==1.0.1
jmespath==0.10.0
lxml==4.5.2
Markdown==3.2.2
more-itertools==8.5.0
packaging==20.4
parse==1.15.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pyparsing==2.4.7
pyquery==1.4.1
pytest==6.0.2
pytest-django==3.9.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
pywebcopy==6.3.0
requests==2.24.0
s3transfer==0.3.3
sentry-sdk==0.17.5
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
termcolor==1.1.0
toml==0.10.1
typing-extensions==3.7.4.3
urllib3==1.25.10
w3lib==1.22.0
wcwidth==0.2.5
